Updating cow apis

+++ src/app/modules/cow/cow.service.ts
@@ -4,7 +4,10 @@ import { paginationHelper } from "../../../shared/paginationHelper";
 import { cowSearchableFields } from "./cow.const";
 import { ICow, ICowFilters } from "./cow.interface";
 import { SortOrder } from "mongoose";
+import { Types } from 'mongoose';
 import { Cow } from "./cow.model";
+import ApiError from "../../../errors/ApiError";
+import httpStatus from "http-status";
 
 const getCows = async (
   filters: ICowFilters,
@@ -88,8 +91,11 @@ const createCow = async (cowData: ICow): Promise<ICow | null> => {
   return newCow;
 };
 
-const getSingleCow = async (id: string): Promise<ICow | null> => {
+const getSingleCow = async (id: string, userId: string): Promise<ICow> => {
   const cow = await Cow.findById(id);
+  if (!cow || !cow.seller.equals(new Types.ObjectId(userId))) {
+    throw new ApiError(httpStatus.UNAUTHORIZED, 'Unauthorized to get this cow');
+  }
   return cow;
 };
 
@@ -103,7 +109,11 @@ const updateCow = async (
   return result;
 };
 
-const deleteCow = async (id: string): Promise<ICow | null> => {
+const deleteCow = async (id: string, userId: string): Promise<ICow | null> => {
+  const cow = await Cow.findById(id);
+  if (!cow || !cow.seller.equals(new Types.ObjectId(userId))) {
+    throw new ApiError(httpStatus.UNAUTHORIZED, 'Unauthorized to delete this cow');
+  }
   const result = await Cow.findByIdAndDelete(id);
   return result;
 };
